user model updated to new model
