+++ apps/roboPortal/urls.py
@@ -2,7 +2,7 @@ from django.contrib import admin
 from django.urls import path, include
 from rest_framework import routers
 # from . import views
-from .serializers import PortalUserViewSet, TeamViewSet, RobothonViewSet 
+# from .serializers import PortalUserViewSet, TeamViewSet, RobothonViewSet
