+++ apps/users/admin.py
@@ -1,9 +1,28 @@
 from django.contrib import admin
-from import_export.admin import ImportExportModelAdmin
-from .models import UserProfile
+from django.contrib.auth.admin import UserAdmin
+from .models import CustomUser
 
-# Register your models here.
+class CustomUserAdmin(UserAdmin):
+    list_display = ['username', 'email', 'auth_provider', 'is_staff', 'is_active', 'date_joined', 'last_login']
+    search_fields = ['username', 'email',]
+    readonly_fields = ['date_joined', 'auth_provider']
+    ordering = ['-date_joined']
 
-@admin.register(UserProfile)
-class UserProfileAdmin(ImportExportModelAdmin):
-    pass
+    filter_horizontal = ()
+    list_filter = ()
+    fieldsets = (
+        (None, {'fields': ('email', 'username', 'password', 'phone_number', 'appick_notification', 'date_joined')}),
+        ('Personal info', {'fields': ('age', 'gender', 'email_verified', 'mobile_verified', 'auth_provider')}),
+        ('Permissions', {
+            'classes': ('collapse',),
+            'fields': ('is_staff', 'is_admin','is_active', 'groups', 'user_permissions')
+            }),
+    )
+    add_fieldsets = (
+        (None, {
+            'classes': ('wide', 'extrapretty'),
+            'fields': ('email', 'username', 'password1', 'password2'),
+        }),
+    )
+
+admin.site.register(CustomUser)
