+++ apps/users/urls.py
@@ -1,19 +1,22 @@
-from rest_framework.routers import DefaultRouter
-from django.urls import path,include
-from .views import UserProfileViewSet,accountView,loginView,registerView, verificationView,forget_pass,forget_pass_email_confirm_view,forgot_change,change
-
-router = DefaultRouter()
-router.register('profiles',UserProfileViewSet,basename='user-profile-viewset')
-app_name = 'users'
+from django.urls import path, include
+from .views import RegisterView, LogoutAPIView, VerifyEmail,  LoginAPIView, PasswordTokenCheckAPI, RequestPasswordResetEmail, UserViewSet
+from rest_framework_simplejwt.views import (
+    TokenRefreshView,
+)
+from rest_framework import routers
+router = routers.DefaultRouter()
+router.register('', UserViewSet)
+app_name = "users"
 
 urlpatterns = [
-     path('',include(router.urls)),
-#     path('account/',accountView,name='account_profile'),
-#     path('verification/',verificationView,name='account_email_verification_sent'),
-#     path('login/',loginView,name='account_login'),
-#     path('register/',registerView,name='account_signup'),
-#     path('forgot/',forget_pass,name='forget-pass'),
-#     path('forgot/confirm/',forget_pass_email_confirm_view,name='forget_pass_email_confirm'),
-#     # path('forgot-change/<str:uid>/<str:token>/',forgot_change, name="forgot_change" ),
-#     path('password-change/',change, name="change"),
+    path('profile/', include(router.urls)),
+    path('register/', RegisterView.as_view(), name="register"),
+    path('login/', LoginAPIView.as_view(), name="login"),
+    path('logout/', LogoutAPIView.as_view(), name="logout"),
+    path('email-verify/', VerifyEmail.as_view(), name='email-verify'),
+    path('token/refresh/', TokenRefreshView.as_view(), name='token_refresh'),
+    path('request-reset-email/', RequestPasswordResetEmail.as_view(),
+         name="request-reset-email"),
+    path('password-reset/<uidb64>/<token>/',
+         PasswordTokenCheckAPI.as_view(), name='password-reset-confirm'),
 ]
